requirements.txt
playwright
Pillow
pytest
pytest-mock
